refactor: overhaul WebSocketHandleResponseMiddleware logic and rename variables for clarity
